refactored css files and added headshots
