+++ components/shared-styled/MemberImage.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const MemberImage = styled.div`
+margin:5px;
+aspect-ratio: 1/1;
+border-radius: 100%;
+overflow: hidden;
+`
+
+export default MemberImage
