+++ components/shared-styled/MemberInfo.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+const MemberInfo = styled.div`
+margin: 20px;
+display: inline-block
+`
+
+export default MemberInfo
