+++ components/shared-styled/Nametag.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const Nametag = styled.div`
+margin:5px;
+
+text-align: center;
+font-weight: bold;
+font-size: 1.1rem
+`
+
+export default Nametag
