+++ components/shared-styled/index.js
@@ -1,3 +1,7 @@
 export { default as SectionTitle } from './SectionTitle'
 export { default as GlobalStyle } from './GlobalStyle'
-export {default as Container} from './Container'
+export { default as Container } from './Container'
+export { default as MemberImage } from './MemberImage'
+export { default as MemberInfo } from './MemberInfo'
+export { default as MembersContainer } from './MembersContainer'
+export { default as Nametag } from './Nametag'
